+++ src/io/ipc/read/array/mod.rs
@@ -1,0 +1,20 @@
+mod primitive;
+pub use primitive::*;
+mod boolean;
+pub use boolean::*;
+mod utf8;
+pub use utf8::*;
+mod binary;
+pub use binary::*;
+mod fixed_size_binary;
+pub use fixed_size_binary::*;
+mod list;
+pub use list::*;
+mod fixed_size_list;
+pub use fixed_size_list::*;
+mod struct_;
+pub use struct_::*;
+mod null;
+pub use null::*;
+mod dictionary;
+pub use dictionary::*;
