+++ src/io/ipc/read/mod.rs
@@ -15,6 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
+mod array;
 mod common;
 mod deserialize;
 mod read_basic;
